Wip refactoring saving function
