+++ src/annotationForm/TaggingTemplate.js
@@ -72,90 +72,75 @@ export default function TaggingTemplate(
   /** Save function * */
   const saveFunction = () => {
     // TODO This code is not DRY, it's the same as in TextCommentTemplate.js
-    if (mediaType !== mediaTypes.AUDIO) {
-      const promises = playerReferences.getCanvases()
-        .map(async (canvas) => {
-          await convertAnnotationStateToBeSaved(annotationState, canvas, windowId);
-          // delete annotationState.maeData.target;
-          return saveAnnotation(annotationState, canvas.id);
-        });
-      Promise.all(promises)
-        .then(() => {
-          closeFormCompanionWindow();
-        });
-    } else {
-      // TODO: Proper save for AUDIO media's annotation
-      console.log('TODO: Proper save for AUDIO media\'s annotation');
-      closeFormCompanionWindow();
-    }
+    saveAnnotation(annotationState);
+  };
 
-    return (
-      <Grid container direction="column" spacing={2}>
-        <Grid item>
-          <Typography variant="formSectionTitle">Tag</Typography>
-        </Grid>
-        <Grid item>
-          <TextField
-            id="outlined-basic"
-            label="Your tag here :"
-            value={annotationState.body.value}
-            variant="outlined"
-            onChange={(event) => updateTaggingValue(event.target.value)}
-          />
-        </Grid>
-        <Grid item>
-          <TargetFormSection
-            currentTime={currentTime}
-            mediaType={mediaType}
-            onChangeTarget={updateTargetState}
-            setCurrentTime={setCurrentTime}
-            setSeekTo={setSeekTo}
-            spatialTarget
-            target={annotationState.maeData.target}
-            timeTarget
-            windowId={windowId}
-            closeFormCompanionWindow={closeFormCompanionWindow}
-            getMediaAudio={getMediaAudio}
-            debugMode={debugMode}
-          />
-        </Grid>
-        <Grid item>
-          <AnnotationFormFooter
-            windowId={windowId}
-            closeFormCompanionWindow={closeFormCompanionWindow}
-            saveAnnotation={saveFunction}
-          />
-        </Grid>
+  return (
+    <Grid container direction="column" spacing={2}>
+      <Grid item>
+        <Typography variant="formSectionTitle">Tag</Typography>
       </Grid>
-    );
-  };
-
-  TaggingTemplate.propTypes = {
-    annotation: PropTypes.shape({
-      adapter: PropTypes.func,
-      body: PropTypes.arrayOf(
-        PropTypes.shape({
-          type: PropTypes.string,
-        }),
-      ),
-      defaults: PropTypes.objectOf(
-        PropTypes.oneOfType(
-          [PropTypes.bool, PropTypes.func, PropTypes.number, PropTypes.string],
-        ),
-      ),
-      drawingState: PropTypes.string,
-      manifestNetwork: PropTypes.string,
-      target: PropTypes.string,
-    }).isRequired,
-    // eslint-disable-next-line react/forbid-prop-types
-    closeFormCompanionWindow: PropTypes.func.isRequired,
-    currentTime: PropTypes.oneOfType([PropTypes.number, PropTypes.instanceOf(null)]).isRequired,
-    // eslint-disable-next-line react/forbid-prop-types
-    getMediaAudio: PropTypes.object.isRequired,
-    // eslint-disable-next-line react/forbid-prop-types
-    saveAnnotation: PropTypes.func.isRequired,
-    setCurrentTime: PropTypes.func.isRequired,
-    setSeekTo: PropTypes.func.isRequired,
-    windowId: PropTypes.string.isRequired,
-  };
+      <Grid item>
+        <TextField
+          id="outlined-basic"
+          label="Your tag here :"
+          value={annotationState.body.value}
+          variant="outlined"
+          onChange={(event) => updateTaggingValue(event.target.value)}
+        />
+      </Grid>
+      <Grid item>
+        <TargetFormSection
+          currentTime={currentTime}
+          mediaType={mediaType}
+          onChangeTarget={updateTargetState}
+          setCurrentTime={setCurrentTime}
+          setSeekTo={setSeekTo}
+          spatialTarget
+          target={annotationState.maeData.target}
+          timeTarget
+          windowId={windowId}
+          closeFormCompanionWindow={closeFormCompanionWindow}
+          getMediaAudio={getMediaAudio}
+          debugMode={debugMode}
+        />
+      </Grid>
+      <Grid item>
+        <AnnotationFormFooter
+          windowId={windowId}
+          closeFormCompanionWindow={closeFormCompanionWindow}
+          saveAnnotation={saveFunction}
+        />
+      </Grid>
+    </Grid>
+  );
 }
+
+TaggingTemplate.propTypes = {
+  annotation: PropTypes.shape({
+    adapter: PropTypes.func,
+    body: PropTypes.arrayOf(
+      PropTypes.shape({
+        type: PropTypes.string,
+      }),
+    ),
+    defaults: PropTypes.objectOf(
+      PropTypes.oneOfType(
+        [PropTypes.bool, PropTypes.func, PropTypes.number, PropTypes.string],
+      ),
+    ),
+    drawingState: PropTypes.string,
+    manifestNetwork: PropTypes.string,
+    target: PropTypes.string,
+  }).isRequired,
+  // eslint-disable-next-line react/forbid-prop-types
+  closeFormCompanionWindow: PropTypes.func.isRequired,
+  currentTime: PropTypes.oneOfType([PropTypes.number, PropTypes.instanceOf(null)]).isRequired,
+  // eslint-disable-next-line react/forbid-prop-types
+  getMediaAudio: PropTypes.object.isRequired,
+  // eslint-disable-next-line react/forbid-prop-types
+  saveAnnotation: PropTypes.func.isRequired,
+  setCurrentTime: PropTypes.func.isRequired,
+  setSeekTo: PropTypes.func.isRequired,
+  windowId: PropTypes.string.isRequired,
+};
